+++ Engine/Controls/UpdateItemControl.xaml
@@ -1,14 +1,15 @@
 ﻿<UserControl
-	x:Class="JocysCom.VS.AiCompanion.Engine.Controls.UpdatesControl"
+	x:Class="JocysCom.VS.AiCompanion.Engine.Controls.UpdateItemControl"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:JcControls="clr-namespace:JocysCom.ClassLibrary.Controls.UpdateControl"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:local="clr-namespace:JocysCom.VS.AiCompanion.Engine.Controls"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	d:DesignHeight="250"
 	x:Name="This"
-	Loaded="This_Loaded"
+	d:DesignHeight="250"
 	d:DesignWidth="800"
+	Loaded="This_Loaded"
 	mc:Ignorable="d">
 	<UserControl.Resources>
 		<ResourceDictionary>
@@ -21,28 +22,8 @@
 			</ResourceDictionary.MergedDictionaries>
 		</ResourceDictionary>
 	</UserControl.Resources>
-	<TabControl
-		x:Name="MainTabControl"
-		Grid.Row="2"
-		Grid.ColumnSpan="2"
-		Margin="4"
-		x:FieldModifier="public"
-		SelectionChanged="MainTabControl_SelectionChanged">
-		<TabItem x:Name="MainTabItem" x:FieldModifier="public">
-			<TabItem.Header>
-				<StackPanel Orientation="Horizontal">
-					<ContentControl Content="{StaticResource Icon_AiCompanion}" />
-					<Label Content="Main" />
-				</StackPanel>
-			</TabItem.Header>
-		</TabItem>
-		<TabItem x:Name="PandocTabItem">
-			<TabItem.Header>
-				<StackPanel Orientation="Horizontal">
-					<ContentControl Content="{StaticResource Icon_Template}" />
-					<Label Content="Pandoc" />
-				</StackPanel>
-			</TabItem.Header>
-		</TabItem>
-	</TabControl>
+	<StackPanel>
+		<JcControls:UpdateTimeUserControl x:Name="UpdateTimePanel" />
+		<JcControls:UpdateUserControl x:Name="UpdatePanel" />
+	</StackPanel>
 </UserControl>
