+++ Engine/Controls/UpdateListControl.xaml
@@ -1,0 +1,49 @@
+﻿<UserControl
+	x:Class="JocysCom.VS.AiCompanion.Engine.Controls.UpdateListControl"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:local="clr-namespace:JocysCom.VS.AiCompanion.Engine.Controls"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	x:Name="This"
+	d:DesignHeight="250"
+	d:DesignWidth="800"
+	Loaded="This_Loaded"
+	mc:Ignorable="d">
+	<UserControl.Resources>
+		<ResourceDictionary>
+			<BooleanToVisibilityConverter x:Key="_BooleanToVisibilityConverter" />
+			<ResourceDictionary.MergedDictionaries>
+				<ResourceDictionary Source="../JocysCom/Controls/Themes/Default.xaml" />
+				<ResourceDictionary Source="../JocysCom/Controls/Themes/Icons.xaml" />
+				<ResourceDictionary Source="../Resources/Icons/Icons_Default.xaml" />
+				<ResourceDictionary Source="../Resources/MainResources.xaml" />
+			</ResourceDictionary.MergedDictionaries>
+		</ResourceDictionary>
+	</UserControl.Resources>
+	<TabControl
+		x:Name="MainTabControl"
+		Grid.Row="2"
+		Grid.ColumnSpan="2"
+		Margin="4"
+		x:FieldModifier="public">
+		<TabItem x:Name="MainTabItem" x:FieldModifier="public">
+			<TabItem.Header>
+				<StackPanel Orientation="Horizontal">
+					<ContentControl Content="{StaticResource Icon_AiCompanion}" />
+					<Label Content="Main" />
+				</StackPanel>
+			</TabItem.Header>
+			<local:UpdateItemControl x:Name="MainUpdateItemPanel" />
+		</TabItem>
+		<TabItem x:Name="PandocTabItem">
+			<TabItem.Header>
+				<StackPanel Orientation="Horizontal">
+					<ContentControl Content="{StaticResource Icon_Template}" />
+					<Label Content="Pandoc" />
+				</StackPanel>
+			</TabItem.Header>
+			<local:UpdateItemControl x:Name="PandocUpdateItemPanel" />
+		</TabItem>
+	</TabControl>
+</UserControl>
